--- banditpylib/learners/collaborative_learner/collaborative_agent_test.py
@@ -1,1 +1,0 @@
-# write test
